@@ -4,14 +4,14 @@
     <groupId>de.zedlitz</groupId>
     <artifactId>phonet4java</artifactId>
     <name>phonet4java</name>
-    <version>1.1</version>
+    <version>1.2</version>
     <packaging>jar</packaging>
     <description>Implementation of various phonetic algorithms in Java.</description>
     <url>https://github.com/jze/phonet4java</url>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.5</maven.compiler.source>
-        <maven.compiler.target>1.5</maven.compiler.target>
+        <maven.compiler.source>7</maven.compiler.source>
+        <maven.compiler.target>7</maven.compiler.target>
     </properties>
     <licenses>
         <license>
